- Added improved DeepClone extension method to try to content with the recent CLR crashes that have been popping up
git-svn-id: https://neurorighter.googlecode.com/svn/trunk@568 471701b0-e017-11dd-9759-139c9ab14fc6

+++ NeuroRighter/SpikeDetection/SpikeDetector.cs
@@ -160,7 +160,7 @@ namespace NeuroRighter.SpikeDetection
                 // Define starting position in current data buffer
                 // Define position in current data buffer
                 int i = numPre + initialSamplesToSkip[channel];
-                int icts  = initialSamplesToSkip[channel].DeepClone();
+                int icts  = (int)initialSamplesToSkip[channel].DeepClone();
                 initialSamplesToSkip[channel] = 0;
 
                 // Create the current data buffer
